requirements.txt
Django>=4.2
djangorestframework>=3.14
Faker>=19.0
black
flake8
